% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{burn_it_down}
\alias{burn_it_down}
\title{Remove all data}
\usage{
burn_it_down(db_path = find_db(), raw_file_dir = find_raw())
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Remove all local ICEWS data artifacts, i.e. local data files and/or database
file.
}
\seealso{
\code{\link[=purge_data_files]{purge_data_files()}}, \code{\link[=purge_db]{purge_db()}}, \code{\link[=remove_db]{remove_db()}}
}
