% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cameo_codes}
\alias{cameo_codes}
\title{CAMEO code dictionary}
\format{Data frame
\describe{
 \item{code}{CAMEO code; the codes need to remain as a character. Coercing
   to integer will create ambiguity, e.g. '010 / Make statement' and
   '10 / DEMAND' will both collapse to 10.}
 \item{name}{CAMEO code name}
 \item{level}{Hierarchy level of this code.}
 \item{lvl0}{Parent root code.}
 \item{lvl1}{Parent root code.}
 \item{description}{Description}
 \item{usage_notes}{Usage notes.}
 \item{example}{Example sentence from which this event type is parsed.}
 \item{order}{Sort order of codes}
}}
\source{
\url{http://eventdata.parusanalytics.com/data.dir/cameo.html}

\url{https://doi.org/10.7910/DVN/28075}
}
\usage{
cameo_codes
}
\description{
CAMEO event type code dictionary
}
\keyword{datasets}
