% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{check_db_exists}
\alias{check_db_exists}
\title{Check and if needed setup database}
\usage{
check_db_exists(db_path, quiet = FALSE)
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\description{
Check and if needed setup database
}
\keyword{internal}
