% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect}
\alias{connect}
\title{Connect to local ICEWS database}
\usage{
connect(db_path = find_db())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\description{
Create a DB connection for the local ICEWS database.
}
