% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to local ICEWS database}
\usage{
connect_to_db(db_path = NULL)
}
\arguments{
\item{db_path}{Path to SQLite database}
}
\description{
Create a DB connection for the local ICEWS database.
}
