% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{Create a local ICEWS database}
\usage{
create_db(db_path = find_db())
}
\arguments{
\item{db_path}{If NULL and the ICEWS_DATA_DIR environment variable is set,
the database will be created at ICEWS_DATA_DIR/db/icews.sqlite3. Otherwise,
an alternative path for the database file.}
}
\description{
Initialize a local ICEWS database.
}
\keyword{internal}
