% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_event_table}
\alias{create_event_table}
\title{Create event table and indices}
\usage{
create_event_table(db_path = find_db())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\description{
Create event table and indices
}
