% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{delete_events}
\alias{delete_events}
\title{Delete events associated with a file}
\usage{
delete_events(file, con = connect())
}
\arguments{
\item{file}{The normalized filename, e.g. "events.1995.[...].tab"}

\item{con}{Connection to the ICEWS database}
}
\description{
Delete events associated with a file
}
