% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\alias{download_icews}
\title{Download ICEWS data files}
\usage{
download_data(to_dir = find_raw(), update = TRUE, dryrun = FALSE)

download_icews(to_dir = find_raw(), update = TRUE, dryrun = FALSE)
}
\arguments{
\item{to_dir}{Path to directory where data files will be downloaded to.}

\item{update}{Update files for which a newer version if available? This will
delete the old file version(s). If FALSE, it will download the new version
but leave the old version in place. There thus will be duplicate event sets.}

\item{dryrun}{Conducts a dry run listing proposed changes, without actually
downloading or deleting anything.}
}
\description{
Download the ICEWS event data from Dataverse
}
\details{
\code{download_data()} will check both ICEWS dataverse repos (both the weekly and
yearly data repos) and download the data files it finds to the location of
the raw data directory (\code{\link[=find_raw]{find_raw()}}). By default, with "update = TRUE", it
will replace existing files with updated versions if they are available, e.g.
if in the raw data directory there is an "events.2017.[date1].tab" file
but on dataverse there is a "events.2017.[date2].tab" file, it will
download the new version and delete the old version once that is done.

Use \code{download_data()} instead of `download_icews()``
}
