% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{download_docs}
\alias{download_docs}
\title{Download documentation}
\usage{
download_docs(docs_dir = find_docs())
}
\arguments{
\item{docs_dir}{Directory containing the documentation files.}
}
\description{
Download documentation files
}
