% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_icews}
\alias{download_icews}
\title{Download ICEWS data files}
\usage{
download_icews(path = Sys.getenv("ICEWS_DATA_DIR"), overwrite = FALSE)
}
\arguments{
\item{path}{Path to directory where data files will be downloaded to.}

\item{overwrite}{Should duplicate files that are already at path be
re-downloaded?}
}
\description{
Download the ICEWS event data from Dataverse
}
