% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_icews.R
\name{dr_icews}
\alias{dr_icews}
\title{What's the current state of things?}
\usage{
dr_icews(db_path = NULL, raw_file_dir = NULL)
}
\arguments{
\item{db_path}{Path to SQLite database file.}

\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Information on the current setup and any data that is present locally.
}
