% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql.R
\name{execute_sql}
\alias{execute_sql}
\title{Execute one of the included SQL definition files at inst/sql}
\usage{
execute_sql(x, db_path)
}
\arguments{
\item{x}{Name of one of the SQL files at inst/sql}

\item{db_path}{Path to SQLite database file.}
}
\description{
Execute one of the included SQL definition files at inst/sql
}
