% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql.R
\name{execute_sql_statements}
\alias{execute_sql_statements}
\title{Execute a vector of SQL statements}
\usage{
execute_sql_statements(x, db_path)
}
\arguments{
\item{x}{A vector containing SQL statements}

\item{db_path}{Path to SQLite database file.}
}
\description{
Execute a vector of SQL statements
}
\keyword{internal}
