% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_path.R
\name{find_path}
\alias{find_path}
\title{Generic path finder}
\usage{
find_path(x, ...)
}
\arguments{
\item{x}{The kind of path to find; "db", "raw", or "docs".}

\item{...}{additional elements to be combined with the file path using
\code{\link[base:file.path]{base::file.path()}}}
}
\description{
Generic path finder
}
\keyword{internal}
