% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{find_raw}
\alias{find_raw}
\title{Find locations}
\usage{
find_raw()
}
\description{
If the optiosn have been set (see [setup_icews()]), `find_raw` will return
the path to the raw data file directory, `find_db` the database file path.
}
\examples{
dontrun{
find_raw()
find_db()
}

}
