% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{format.icews_opts}
\alias{format.icews_opts}
\title{Format ICEWS opts}
\usage{
\method{format}{icews_opts}(x, ...)
}
\arguments{
\item{x}{Object generated by \code{\link[=get_icews_opts]{get_icews_opts()}}.}

\item{...}{Not used}
}
\description{
Format ICEWS opts
}
