% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{format.icews_plan}
\alias{format.icews_plan}
\title{Format method for plan}
\usage{
\method{format}{icews_plan}(x, ...)
}
\arguments{
\item{x}{An object of class "icews_plan"}

\item{...}{Not used}
}
\description{
Format method for plan
}
\keyword{internal}
