% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_db_state}
\alias{get_db_state}
\title{Get database state}
\usage{
get_db_state(db_path = find_db())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\description{
Determine what data files, event sets, and version are currently in the
database.
}
