% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{get_doi}
\alias{get_doi}
\title{Get ICEWS dataverse DOI}
\usage{
get_doi()
}
\value{
A list containing the historic and weekly DOIs.
}
\description{
Contains the Harvard Dataverse DOIs for ICEWS
}
\examples{
get_doi()

}
