% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_dataverse.R
\name{get_dvn_manifest}
\alias{get_dvn_manifest}
\title{List dataverse files}
\usage{
get_dvn_manifest(
  icews_doi = get_doi(),
  server = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{icews_doi}{DOI of the main ICEWS repo on Dataverse, see \code{\link[=get_doi]{get_doi()}}}

\item{server}{For unit tests only; default is set to \code{\link[dataverse:get_dataset]{dataverse::get_dataset()}} default.}
}
\value{
A list of length 2, containing:
\itemize{
\item file_list: a summary list of all files on DVN, consisting of the data files
but also documentation and metadata files.
\item dataverse_dataset: a list tibble with two columns:
\itemize{
\item \code{repo}: "weekly" or "historic"`
\item \code{content}: objects of class "dataverse_dataset", returned by
\code{\link[dataverse:get_dataset]{dataverse::get_dataset()}}.
}
}
}
\description{
Get a listing of files on dataverse
}
