% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_icews_opts}
\alias{get_icews_opts}
\title{Get ICEWS options}
\usage{
get_icews_opts()
}
\description{
List the current ICEWS-related option values
}
\seealso{
\code{\link[=set_icews_opts]{set_icews_opts()}}, \code{\link[=unset_icews_opts]{unset_icews_opts()}}, \code{\link[=setup_icews]{setup_icews()}}
}
