% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_local_state}
\alias{get_local_state}
\title{Get local file state}
\usage{
get_local_state(raw_file_dir = find_raw())
}
\arguments{
\item{raw_file_dir}{Directory containing raw data files}
}
\description{
Determine what data files are present locally, and which event sets and
versions they contain.
}
