% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{goldstein_mappings}
\alias{goldstein_mappings}
\title{CAMEO event Goldstein scores}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 312 rows and 5 columns.
}
\source{
\url{http://eventdata.parusanalytics.com/data.dir/cameo.html}

\doi{10.7910/DVN/28075}
}
\usage{
goldstein_mappings
}
\description{
List mapping CAMEO codes to Goldstein scores
}
\examples{
data(goldstein_mappings)
head(goldstein_mappings)
}
\keyword{datasets}
