% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icews-package.R
\docType{package}
\name{icews}
\alias{icews}
\alias{icews-package}
\title{Get ICEWS Event Data}
\description{
Download utilities and other helpers for getting the ICEWS event
}
\references{
Boschee, Elizabeth; Lautenschlager, Jennifer; O'Brien, Sean;
      Shellman, Steve; Starz, James; Ward, Michael, 2015, ``ICEWS Coded
      Event Data'', \url{https://doi.org/10.7910/DVN/28075}, Harvard Dataverse.
}
