% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{icews_sample}
\alias{icews_sample}
\title{ICEWS event data sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4993 rows and 22 columns.
}
\source{
\doi{10.7910/DVN/28075}
}
\usage{
icews_sample
}
\description{
A sample of the ICEWS event data consisting of the 4,993 events between 1 and
3 June 2018.
}
\examples{
data("icews_sample")
str(icews_sample)
}
\keyword{datasets}
