% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{icews_to_gwcode}
\alias{icews_to_gwcode}
\title{Translate country to G&W country codes}
\usage{
icews_to_gwcode(country, date)
}
\arguments{
\item{country}{The character country name in ICEWS}

\item{date}{Event date; some countries like Yugoslavia/Serbia and Montenegro/Serbia
change over time.}
}
\description{
Translate country names to Gleditsch & Ward country codes.
}
\examples{
# Territories are correctly associated with metropole
df <- dplyr::tibble(
  country = c("United States", "Puerto Rico", "Guam"),
  event_date = rep(as.Date("2018-01-01")),
  stringsAsFactors = FALSE)
df$gwcode <- icews_to_gwcode(df$country, df$event_date)
df

# Time-varying translation
df <- dplyr::tibble(
  country = c("Serbia", "Serbia"),
  event_date = as.Date(c("2006-06-04", "2006-06-05")),
  stringsAsFactors = FALSE)
df$gwcode <- icews_to_gwcode(df$country, df$event_date)
df
}
