% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ingest_from_file}
\alias{ingest_from_file}
\title{Ingest a raw events file}
\usage{
ingest_from_file(raw_file_path = NULL, db_path = find_db())
}
\arguments{
\item{raw_file_path}{Path to a TSV file with events.}

\item{db_path}{Path to SQLite database}
}
\description{
Ingest a raw data file into the database
}
\keyword{internal}
