% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ingest_raw_data_file}
\alias{ingest_raw_data_file}
\title{Ingest a raw events file}
\usage{
ingest_raw_data_file(raw_file_path, db_path)
}
\arguments{
\item{raw_file_path}{Directory containing the raw event TSV files.}

\item{db_path}{Path to SQLite database}
}
\description{
Ingest a raw data file into the database
}
