% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_indices}
\alias{list_indices}
\title{List indices in database}
\usage{
list_indices(db_path = find_db())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\description{
List the indices currently in the "event" table in the database.
}
\keyword{internal}
