% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_local_files.R
\name{list_local_files}
\alias{list_local_files}
\title{List the raw data file paths}
\usage{
list_local_files(raw_file_dir = find_raw(), full_names = TRUE)
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files}

\item{full_names}{Return the full path or only file name?}
}
\description{
Get a list of the paths for any local raw data ".tab" files.
}
