% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_state.R
\name{list_source_files}
\alias{list_source_files}
\title{List of ingested source files}
\usage{
list_source_files(db_path = find_db())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\value{
A character vector of source file names.
}
\description{
List the source files from which events have been ingested to the DB already.
}
