% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events_tsv.R
\name{normalize_column_names}
\alias{normalize_column_names}
\title{Normalize column names}
\usage{
normalize_column_names(x)
}
\arguments{
\item{x}{Tibble containing raw events}
}
\description{
Normalize column names
}
