% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_dataverse.R
\name{normalize_label}
\alias{normalize_label}
\title{Normalize a data file label}
\usage{
normalize_label(x)
}
\arguments{
\item{x}{character string, DVN file label}
}
\value{
A character vector with normalized file names.
}
\description{
Normalize the dataverse file label, i.e. substitute ".tab" for zipped files
and alter duplicate file names
}
\keyword{internal}
