% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_paths}
\alias{normalize_paths}
\title{Normalize paths}
\usage{
normalize_paths(db_path, raw_file_dir)
}
\arguments{
\item{db_path}{Location of database SQLite file}

\item{raw_file_dir}{Directory for raw data files}
}
\description{
This takes care of finding paths when the environment variable is set and
paths are at the default NULL values.
}
