% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{optimize_db}
\alias{optimize_db}
\title{Optimize database}
\usage{
optimize_db(db_path, vacuum = TRUE, optimize = TRUE)
}
\arguments{
\item{db_path}{Path to database file}

\item{optimize}{Call "PRAGMA optimize"?}

\item{vaccum}{Call "VACUUM" command?}
}
\description{
Call SQLite's database optimizer and vacuum
}
