% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-helpers.R
\name{parse_dataset}
\alias{parse_dataset}
\title{Identify dataset contained in file}
\usage{
parse_dataset(x)
}
\arguments{
\item{x}{a normalized file name}
}
\description{
Identify which time period is nominally covered by a file. This is kept
around from prior version of the package, in case it becomes useful again.
E.g. to allow for time range specific downloading.
}
\keyword{internal}
