% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{parse_label}
\alias{parse_label}
\title{Parse a data file label}
\usage{
parse_label(label)
}
\arguments{
\item{label}{character string}
}
\value{
A data frame with components:
\itemize{
\item label
\item file
\item is_date: TRUE or FALSE
\item dataset
\item version
}
}
\description{
Parse the Dataverse file label to identify whether it contains data and if so which
event set and file version.
}
\keyword{internal}
