% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_database_changes}
\alias{plan_database_changes}
\title{Plan database changes}
\usage{
plan_database_changes(db_path = find_db(), raw_file_dir = find_raw(),
  keep_files = getOption("icews.keep_files"), use_local = TRUE)
}
\arguments{
\item{db_path}{Path to SQLite database file}

\item{raw_file_dir}{Path to raw data file directory.}

\item{keep_files}{Retain raw data files?}

\item{use_local}{Use local files if available?}
}
\description{
Plan database changes related to download/updating
}
