% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_database_sync}
\alias{plan_database_sync}
\title{Plan database sync}
\usage{
plan_database_sync(db_path = find_db(), raw_file_dir = find_raw())
}
\arguments{
\item{db_path}{Path to SQLite database file}

\item{raw_file_dir}{Path to raw data file directory.}
}
\description{
Plan database synchronization with local files, without downloading any
new files
}
\keyword{internal}
