% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_file_changes}
\alias{plan_file_changes}
\title{Plan file changes}
\usage{
plan_file_changes(raw_file_dir = find_raw())
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files}
}
\description{
Plan file changes related to download/updating
}
\keyword{internal}
