% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.icews_opts}
\alias{print.icews_opts}
\title{Print current ICEWS options}
\usage{
\method{print}{icews_opts}(x, ...)
}
\arguments{
\item{x}{Object generated by \code{\link[=get_icews_opts]{get_icews_opts()}}.}

\item{...}{Not used}
}
\description{
Print current ICEWS options
}
