% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{print.icews_plan}
\alias{print.icews_plan}
\title{Print method for plan}
\usage{
\method{print}{icews_plan}(x, ...)
}
\arguments{
\item{x}{A tibble containing the plan output}

\item{...}{Not used}
}
\description{
Print method for plan
}
