% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{print_plan}
\alias{print_plan}
\title{Print a summary of a plan}
\usage{
print_plan(plan)
}
\arguments{
\item{plan}{A tibble containing the plan output}
}
\description{
Summarize the actions to be performed in a plan
}
