% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{purge_data_files}
\alias{purge_data_files}
\title{Purge file downloads}
\usage{
purge_data_files(raw_file_dir = find_raw())
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files}
}
\description{
Removes the downloaded raw event TSV files
}
\seealso{
[purge_db()], [remove_db()], [burn_it_down()]
}
