% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(query, db_path = find_db())
}
\arguments{
\item{query}{SQL query string}

\item{db_path}{Path to SQLite database file}
}
\description{
Get results from a query to the database
}
