% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_events_tsv}
\alias{read_events_tsv}
\title{Read events TSV}
\usage{
read_events_tsv(file, ...)
}
\arguments{
\item{file}{Path to a raw events tab-delimited file (".tab").}

\item{...}{Other options passed to [readr::read_tsv()].}
}
\description{
Wrapper around [readr::read_tsv()] to read in the raw events tab-delimited
files. It provides the correct column type specification and disables the
default quote escape option, as the files already have escaped quotes.
}
