% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_icews}
\alias{read_icews}
\title{Read and combine raw data files}
\usage{
read_icews(path = Sys.getenv("ICEWS_DATA_DIR"))
}
\arguments{
\item{path}{Path to directory where data files will be downloaded to.}
}
\description{
Read the entire ICEWS event data into memory. This takes up several (2-3 in 2018) GB.
}
