% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{reconcile_state}
\alias{reconcile_state}
\title{Reconcile state}
\usage{
reconcile_state(path = NULL)
}
\arguments{
\item{path}{Directory containing the raw event TSV files.}
}
\description{
Reconcile the local and dataverse file state
}
