% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{remove_db}
\alias{remove_db}
\title{Remove database file}
\usage{
remove_db(db_path, directory = FALSE)
}
\arguments{
\item{db_path}{Path to SQLite database file.}

\item{directory}{Should the directory be removed also? This will delete any
other files that are in it.}
}
\description{
Removes the SQLite database file. This is quicker than deleting the events
table.
}
\seealso{
\code{\link[=purge_db]{purge_db()}}, \code{\link[=delete_events]{delete_events()}}
}
