% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{remove_file}
\alias{remove_file}
\title{Remove a data file}
\usage{
remove_file(raw_file_path)
}
\arguments{
\item{raw_file_path}{Data file to be removed}
}
\description{
Remove a data file
}
\keyword{internal}
