% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_icews_opts}
\alias{set_icews_opts}
\title{Set ICEWS options}
\usage{
set_icews_opts(data_dir, use_db, keep_files)
}
\arguments{
\item{data_dir}{Where should the raw TSV data be kept?}

\item{use_db}{Store events in a SQLite database?}

\item{keep_files}{keep_files If using a database, retain raw data TSV files?}
}
\description{
Set ICEWS options
}
