% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_database.R, R/state_dataverse.R,
%   R/state_local_files.R
\name{get_db_state}
\alias{get_db_state}
\alias{get_dvn_state}
\alias{get_local_state}
\title{Get DVN/local file/database state}
\usage{
get_db_state(db_path = find_db())

get_dvn_state(icews_doi = get_doi(), server = Sys.getenv("DATAVERSE_SERVER"))

get_local_state(raw_file_dir = find_raw())
}
\arguments{
\item{db_path}{Path to SQLite database files.}

\item{icews_doi}{DOI of the main ICEWS repo on Dataverse, see \code{\link[=get_doi]{get_doi()}}}

\item{server}{For unit tests only; default is set to \code{\link[dataverse:get_dataset]{dataverse::get_dataset()}} default.}

\item{raw_file_dir}{Directory containing raw data files}
}
\value{
For \code{get_dvn_manifest}, a tibble with the following columns:
\itemize{
\item dvn_repo: "historic" or "weekly", see \code{\link[=get_doi]{get_doi()}}
\item dvn_file_label: the file label on dataverse, possibly non-unique
\item dvn_file_id: the integer file ID on dataverse
\item file_name: the normalized, unique file name, see \code{\link[=normalize_label]{normalize_label()}}
}

For \code{get_local_state} and \code{get_db_state}, a tibble with columns:
\itemize{
\item file_name: the full source data file name, e.g.
"events.1995.20150313082510.tab"; see \code{\link[=normalize_label]{normalize_label()}}
}
}
\description{
Determine what data files are currently on dataverse, in the local files,
or in the local database.
}
\details{
The data files (tab-separated files, ".tab") on dataverse that
contain the raw event data follow a common format denoting the set of
events contained in a file and which version of the event data and/or file
dump they correspond to. For example, "events.1995.20150313082510.tab"
contains events for 1995 and the version is denoted by the timestamp,
"20150313082510".

The download and update functions
(\code{\link[=update_icews]{update_icews()}}, \code{\link[=download_data]{download_data()}}) will recognize which event sets
are locally available or still need to be downloaded, and whether any
local even sets have been superseded by a new version in dataverse, by
using
}
\examples{
\dontrun{
# Remote (DVN) state
get_dvn_state()
# Local file state
get_local_state()
# Database state
get_db_state()
}

}
