% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sync_db}
\alias{sync_db}
\title{Synchronize/update local ICEWS database}
\usage{
sync_db(db_path = NULL, raw_file_dir = NULL)
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Maintain a current set of ICEWS events in the local database that match
the latest versions on DVN. If needed, create the database and download data
files.
}
