% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sync_db_with_files}
\alias{sync_db_with_files}
\title{Sync database and raw files}
\usage{
sync_db_with_files(db_path = NULL, raw_file_dir = NULL,
  dryrun = FALSE)
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{raw_file_dir}{Directory containing the raw event TSV files.}

\item{dryrun}{List file changes to be performed without taking any action.}
}
\description{
Synchronize database contents with local raw data files.
}
