% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{unset_icews_opts}
\alias{unset_icews_opts}
\title{Unset ICEWS options}
\usage{
unset_icews_opts()
}
\value{
Silently returns a list of prior option values to make it possible
reset the values if needed. See \code{\link[=get_icews_opts]{get_icews_opts()}} and
\code{\link[=set_icews_opts]{set_icews_opts()}}.
}
\description{
Unset ICEWS options
}
