% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update}
\alias{update}
\title{Update database and files}
\usage{
update(dryrun = FALSE, use_db = getOption("icews.use_db"),
  keep_files = getOption("icews.keep_files"), db_path = find_db(),
  raw_file_dir = find_raw())
}
\arguments{
\item{dryrun}{Just list changes that would be made, without making them.}

\item{use_db}{Store events in a SQLite database?}

\item{keep_files}{If using a database, retain raw data TSV files?}

\item{db_path}{Path to SQLite database file}

\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Maintain a current set of ICEWS events in the local database that match
the latest versions on DVN. If needed, create the database and download data
files.
}
