% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_data_to_db}
\alias{write_data_to_db}
\title{Augment and write events to DB}
\usage{
write_data_to_db(events, file, db_path = find_db())
}
\arguments{
\item{events}{data.frame containing ICEWS events}

\item{file}{Name of the TSV source file from which events came}

\item{db_path}{Path to SQLite database file}
}
\description{
Adds some columns to events before writing to DB.
}
\keyword{internal}
