% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_event_table}
\alias{create_event_table}
\title{Create event table and indices}
\usage{
create_event_table(db_path)
}
\arguments{
\item{db_path}{Path to SQLite database file.}
}
\description{
Create event table and indices
}
\keyword{internal}
