% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\alias{download_icews}
\title{Download ICEWS data files}
\usage{
download_data(to_dir = find_raw(), update = TRUE, dryrun = FALSE)

download_icews(to_dir = find_raw(), update = TRUE, dryrun = FALSE)
}
\arguments{
\item{to_dir}{Path to directory where data files will be downloaded to.}

\item{update}{Update files for which a newer version if available? This will
delete the old file version(s). If FALSE, it will download the new version
but leave the old version in place. There thus will be duplicate event sets.}

\item{dryrun}{Conducts a dry run listing proposed changes, without actually
downloading or deleting anything.}
}
\description{
Download the ICEWS event data from Dataverse
}
\details{


Use \code{download_data()} instead of `download_icews()``
}
