% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\title{Download a single file from Dataverse}
\usage{
download_file(file, to_dir, repo = "historic", file_id = NULL, new_name = NULL)
}
\arguments{
\item{file}{Name (label) of the file on DVN to download. If zipped, it will
automatically be unzipped.}

\item{to_dir}{Destination directory.}

\item{repo}{Which repo is the file in? ("historic" or "daily")}

\item{file_id}{Optionally, integer file ID. If this is specified, it will
preferentially be used over the file name to download the file. This is
needed for duplicate file names.}

\item{new_name}{Optionally, a new file name for the file. Useful for
duplicate files.}
}
\value{
Silently returns the downloaded file name.
}
\description{
Download a single file from the ICEWS Dataverse repository and unzip if neccessary
}
\details{
To get a list of files available for download, see \link{get_dvn_manifest}.
}
