% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\title{Download a single file from DVN}
\usage{
download_file(file, to_dir, repo = "historic")
}
\arguments{
\item{file}{Name of the file on DVN to download. If zipped.}

\item{to_dir}{Destination directory.}

\item{repo}{Which repo is the file in? ("historic" or "daily")}
}
\value{
Silently returns the downloaded file name.
}
\description{
Download a single file from the ICEWS DVN repository and unzip if neccessary
}
\details{
To get a list of files available for download, see \link{get_dvn_manifest}.
}
