% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icews-package.R
\docType{package}
\name{icews}
\alias{icews}
\alias{icews-package}
\title{Get ICEWS Event Data}
\description{
Download utilities and other helpers for getting the ICEWS event
data from Dataverse.
}
\section{Package options}{


Icews uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{icews.debug}: set to TRUE to print additional diagnostic messages
in various functions.
}

Options that control file and database paths, and which should be set
with \code{\link[=setup_icews]{setup_icews()}}:

\itemize{
\item \code{icews.data_dir}: Path to the directory in which data will be kept
locally. Depending on the other settings, subdirectories like "raw" and
"db" will be created within.
\item \code{icews.use_db}: Use a SQLite database to store the data?
\item \code{icews.keep_files}: If "use_db" is TRUE, retain the raw data TSV
files or delete after ingesting to the database?
}

See \url{https://github.com/andybega/icews} and
\url{https://www.andybeger.com/icews/} for
more information.
}

\references{
Boschee, Elizabeth; Lautenschlager, Jennifer; O'Brien, Sean;
Shellman, Steve; Starz, James; Ward, Michael, 2015, ``ICEWS Coded
Event Data'', \url{https://doi.org/10.7910/DVN/28075},
Harvard Dataverse.

Boschee, Elizabeth; Lautenschlager, Jennifer; O'Brien, Sean; Shellman,
Steve; Starz, James, 2018, ``ICEWS Weekly Event Data'',
\url{https://doi.org/10.7910/DVN/QI2T9A},
Harvard Dataverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.andybeger.com/icews/}
  \item \url{https://github.com/andybega/icews}
  \item Report bugs at \url{https://github.com/andybega/icews/issues}
}

}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

}
