% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ingest_from_memory}
\alias{ingest_from_memory}
\title{Ingest a file without retaining}
\usage{
ingest_from_memory(file, con = connect())
}
\arguments{
\item{file}{The normalized filename, e.g. "events.1995.[...].tab"}

\item{con}{Connection to the ICEWS database}
}
\description{
The "in memory" part is not working because I haven't figured out how to
unzip a raw vector in memory.
}
