% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_icews}
\alias{read_icews}
\title{Read ICEWS events into memory}
\usage{
read_icews(path = NULL)
}
\arguments{
\item{path}{Either path to SQLite database file or raw file directory. If
NULL (default), the global options will be used instead.}
}
\description{
Read the entire ICEWS event data into memory. This takes up several (2-3 in 2018) GB.
}
\seealso{
\code{\link[=query_icews]{query_icews()}}, \code{\link[=read_events_tsv]{read_events_tsv()}}
}
