% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icews.R
\name{read_icews}
\alias{read_icews}
\title{Read ICEWS events into memory}
\usage{
read_icews(path = NULL, n_max = NULL)
}
\arguments{
\item{path}{Either path to SQLite database file or raw file directory. If
NULL (default), the global options will be used instead.}

\item{n_max}{Maximum number of events to load, if NULL (default), no limit is
used, i.e. all events are imported.}
}
\description{
Read the entire ICEWS event data into memory. This takes up several GB;
approximately 3GB in mid-2020.
}
\examples{
\dontrun{
# Both of these work, but reading from files should be quicker
from_files <- read_icews(find_raw())
from_db    <- read_icews(find_db())
}

}
\seealso{
\code{\link[=query_icews]{query_icews()}}, \code{\link[=read_events_tsv]{read_events_tsv()}}
}
