% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_icews}
\alias{read_icews}
\title{Read and combine raw data files}
\usage{
read_icews(raw_file_dir = find_raw())
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Read the entire ICEWS event data into memory. This takes up several (2-3 in 2018) GB.
}
