% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_icews_raw}
\alias{read_icews_raw}
\title{Read data from raw files}
\usage{
read_icews_raw(raw_file_dir)
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files.}
}
\description{
Read data from raw files
}
