% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql.R
\name{read_sql_statements}
\alias{read_sql_statements}
\title{Read one of the included SQL files at inst/sql and split it into a vector
of separate SQL statements}
\usage{
read_sql_statements(x)
}
\arguments{
\item{x}{Path to a .sql file containing SQL statements (not queries)}
}
\description{
Read one of the included SQL files at inst/sql and split it into a vector
of separate SQL statements
}
\keyword{internal}
