% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_icews_opts}
\alias{set_icews_opts}
\title{Set ICEWS option values}
\usage{
set_icews_opts(x, ...)
}
\arguments{
\item{x}{Either an object of class "icews_opts" generated by \code{\link[=get_icews_opts]{get_icews_opts()}}
or \code{\link[=unset_icews_opts]{unset_icews_opts()}}, or a string for the data directory path.}

\item{...}{arguments for other methods}
}
\description{
Set ICEWS option values
}
\seealso{
\code{\link[=setup_icews]{setup_icews()}}
}
