% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_icews}
\alias{setup_icews}
\title{Setup ICEWS environment variables}
\usage{
setup_icews(data_dir, use_db, r_environ = FALSE)
}
\arguments{
\item{data_dir}{Where should the raw TSV data be kept?}

\item{use_db}{Use a SQLite database to store and work with the data?}

\item{r_environ}{If TRUE, this will write config parameters to a .Renviron file.}
}
\description{
Checks for and creates several ICEWS environment variables.
}
