% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_dvn_manifest}
\alias{get_dvn_manifest}
\alias{get_local_state}
\alias{get_db_state}
\title{Get DVN/local file/database state}
\usage{
get_dvn_manifest()

get_local_state(raw_file_dir = find_raw())

get_db_state(db_path = find_db())
}
\arguments{
\item{raw_file_dir}{Directory containing raw data files}

\item{db_path}{Path to SQLite database files.}
}
\value{
For \code{get_local_state} and \code{get_db_state}, a tibble with columns:
\itemize{
\item db/local_file: the full source data file name, e.g. "events.1995.20150313082510.tab".
\item db/local_data_set: the event set contained in the file, e.g. "events.1995".
\item db/local_version: the version of the file/event set, e.g. "20150313082510"
}

For \code{get_dvn_manifest}, a list of length 3, containing:
\itemize{
\item data_files: a tibble similar to those returned by \code{get_local_state} and
\code{get_db_state}, with an additional column for the file label on dataverse
since some files are zipped, i.e. ending with ".tab.zip" instead of ".tab".
\item files: a summary list of all files on DVN, consisting of the data files
but also documentation and metadata files.
\item dataverse_dataset: an object of class "dataverse_dataset", returned by
\code{\link[dataverse:get_dataset]{dataverse::get_dataset()}}.
}
}
\description{
Determine what data files, event sets, and version are currently on dataverse,
in the local files, or in the local database.
}
\details{
The data files (tab-separated files, ".tab") on dataverse that
contain the raw event data follow a common format denoting the set of
events contained in a file and which version of the event data and/or file
dump they correspond to. For example, "events.1995.20150313082510.tab"
contains events for 1995 and the version is denoted by the timestamp,
"20150313082510".

The download and update functions
(\code{\link[=update_icews]{update_icews()}}, \code{\link[=download_data]{download_data()}}) will recognize which event sets
are locally available or still need to be downloaded, and whether any
local even sets have been superseded by a new version in dataverse, by
using
}
\examples{
\dontrun{
# Remote (DVN) state
get_dvn_manifest()
# Local file state
get_local_state()
# Database state
get_db_state()
}

}
