% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sync_db_with_files}
\alias{sync_db_with_files}
\title{Synchronize DB with raw files}
\usage{
sync_db_with_files(
  raw_file_dir = find_raw(),
  db_path = find_db(),
  dryrun = FALSE
)
}
\arguments{
\item{raw_file_dir}{Directory containing the raw event TSV files.}

\item{db_path}{Path to SQLite database file}

\item{dryrun}{List changes to be performed without taking any action.}
}
\description{
Synchronize DB with any local files found, without downloading new files.
}
